CC-ML/requirements.txt
mysql-connector-python
pandas
astor
tensorflow==2.15.1
numpy
Flask
PyJWT
google-cloud-storage
